social-syrup/phase1_x_upload/requirements.txt
requests==2.32.3
requests-oauthlib==2.0.0
python-dotenv==1.0.1
Pillow==10.4.0
